Add files via upload
The idea here is to consider prime factors of a factorial of N. A trailing zero will always be produced by 2&5. Hence, count the number of 2's and 5's in the factorial. But since 2 comes before 5, there is always a case such that number of 2's is greater than or equal to 5.
Hence we should count the number of 5's. But in case of 25, 125 etc there are more number of 5's. Hence keep dividing the number by multiples of 5 to get the exact count of trailing zeros.